split cart js module
